refactor: 归位音频与预览领域实现
